requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.0.11
Django==1.10.6
enum34==1.1.6
functools32==3.2.3.post2
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
matplotlib==2.0.0
networkx==1.11
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
scandir==1.5
scikit-image==0.13.0
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.7
traitlets==4.3.2
wcwidth==0.1.7
